requirements.txt
torch>=2.1.0
transformers>=4.38.0
datasets>=2.16.0
peft>=0.8.0
trl>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
scipy>=1.11.0
numpy>=1.24.0
pyyaml>=6.0
sentencepiece>=0.1.99
protobuf>=3.20.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
